sprints/sprint-1-llm-foundations/week-3/Project/Mini-RAG PoC/requirements.txt
llama-index-core==0.10.35
llama-index-llms-openai==0.1.9
llama-index-vector-stores-pinecone==0.1.3
llama-index-retrievers-bm25==0.1.3
llama-index-embeddings-openai==0.1.5
llama-index-readers-file==0.1.6
pandas
openai

